Fixing fastjet mode for pp Updating jet resolution study

+++ test/submitMarlinRun.py
@@ -21,20 +21,29 @@ os.system('cmsMkdir %s'%opt.output)
 outDir=os.path.expandvars(opt.locoutput)
 os.system('mkdir -p %s'%outDir)
 
+isCastor=False
+lsCommand='cmsLs %s | awk \'{print $5}\''%opt.input
+if opt.input.find('castor/cern.ch')>=0 :
+    isCastor=True
+    lsCommand='rfdir %s | awk \'{print $9}\''%opt.input
 
 j=0
-fList=commands.getstatusoutput('rfdir %s | awk \'{print $9}\''%opt.input)[1].split('\n')
+fList=commands.getstatusoutput(lsCommand)[1].split('\n')
 for f in fList:
+
+    #get the base name of the file (normalize castor and EOS outputs)
+    f=os.path.basename(f)
+    if f.find('.slcio')<0 : continue
+
     j=j+1
     jobName='ilcsim%d'%j
-
-    #create the steer for the job
-
+    
     #create a standalone Marlin job
     scriptFile = open('%s/runJob_%d.sh'%(outDir,j), 'w')
     scriptFile.write('#!/bin/bash\n')
     scriptFile.write('source /afs/cern.ch/eng/clic/software/x86_64-slc5-gcc41/ILCSOFT/v01-16/init_ilcsoft.sh\n')
-    scriptFile.write('rfcp %s/%s /tmp/%s \n'%(opt.input,f,f))
+    if isCastor : scriptFile.write('rfcp %s/%s /tmp/%s \n'%(opt.input,f,f))
+    else        : scriptFile.write('cmsStage %s/%s /tmp/%s \n'%(opt.input,f,f))
     scriptFile.write('Marlin %s/../%s --MyLCIOOutputProcessor.LCIOOutputFile=/tmp/fj_%s --global.LCIOInputFiles=/tmp/%s \n'%(outDir,opt.steer,f,f))
     scriptFile.write('cmsStage /tmp/fj_%s %s \n'%(f,opt.output))
     scriptFile.write('rm /tmp/*.slcio \n')
@@ -43,4 +52,4 @@ for f in fList:
 
     os.system('chmod u+rwx %s/runJob_%d.sh'%(outDir,j))
     os.system("bsub -q %s -R \"%s\" -J %s \'%s/runJob_%d.sh\'"%(opt.queue,opt.requirementtoBatch,jobName,outDir,j))
-
+
